move ClusteredVariantAccessioningService out of the non-human package since this is used for human and nonhuman
